+++ helpers/helpers.go
@@ -1,68 +1,12 @@
 package helpers
 
 import (
-	"database/sql"
 	"encoding/json"
-	"errors"
 	"net/http"
-
-	"github.com/ndv6/tsaving/database"
 )
 
 //untuk ngehandle error"
 func HTTPError(w http.ResponseWriter, status int, errorMessage string) {
 	w.WriteHeader(status)
 	json.NewEncoder(w).Encode(map[string]string{"error": errorMessage})
-}
-
-//untuk ngecek input rekening apakah benar atau tidak.
-func CheckAccountVA(db *sql.DB, VaNum string, id int) (err error) {
-	var exist bool
-	err = db.QueryRow("SELECT EXISTS(SELECT va_num FROM virtual_accounts INNER JOIN customers ON virtual_accounts.account_num = customers.account_num WHERE va_num = $1 AND cust_id = $2)", VaNum, id).Scan(&exist)
-	if err != nil {
-		return
-	}
-	if !exist {
-		err = errors.New("invalid virtual account number")
-		return
-	}
-	return
-}
-
-func CheckAccount(db *sql.DB, AccountNum string, id int) (err error) {
-	// AccountNumber := 0
-	var exist bool
-	err = db.QueryRow("SELECT EXISTS(SELECT account_num FROM customers WHERE account_num = $1 AND cust_id = $2)", AccountNum, id).Scan(&exist)
-	if err != nil {
-		return
-	}
-	if !exist {
-		err = errors.New("invalid account number")
-		return
-	}
-	return
-}
-
-func CheckBalance(target string, accNumber string, amount int, db *sql.DB) (status bool) {
-	if target == "MAIN" {
-		sourceBalance, err := database.GetBalanceAcc(accNumber, db)
-		if err != nil {
-			return
-		}
-		if sourceBalance.AccountBalance < amount || amount <= 0 {
-			return
-		}
-		status = true
-	}
-	if target == "VA" {
-		sourceBalance, err := database.GetBalanceVA(accNumber, db)
-		if err != nil {
-			return
-		}
-		if sourceBalance.VaBalance < amount || amount <= 0 {
-			return
-		}
-		status = true
-	}
-	return
 }
