+++ test/virtual_accounts/virtual_accounts.go
@@ -5,16 +5,11 @@ import (
 	"time"
 )
 
-type Event struct {
-	Name      string
-	StartDate time.Time
-	EndDate   time.Time
-}
 
-func TestActiveEvents(t *testing.T) {
+func TestVacToMain(t *testing.T) {
 
 	// ini dibikin skenario, kalau inputnya begini, dan hasilnya oke, nanti akan keluar oke.
-	res := ActiveEvents(le, time.Date(2020, 07, 03, 0, 0, 0, 0, time.UTC))
+	res := (le, time.Date(2020, 07, 03, 0, 0, 0, 0, time.UTC))
 	if len(res) != 2 {
 		t.Fatalf("Expect 2 Events, got: %v", len(res))
 	}
